AP_HAL_F4Light: fixed OSD compilation issue
